[codeview] Use one byte for S_FRAMECOOKIE CookieKind and add flags byte
We bailed out while printing codeview for an MSVC compiled
SemaExprCXX.cpp that used this record. The MS reference headers look
incorrect here, which is probably why we had this bug. They use a 32-bit
enum as the field type, but the actual record appears to use one byte
for the cookie kind followed by a flags byte.

llvm-svn: 273691

+++ llvm/include/llvm/DebugInfo/CodeView/CodeView.h
@@ -465,7 +465,7 @@ enum class BinaryAnnotationsOpCode : uint32_t {
 };
 
 // Corresponds to CV_cookietype_e enum.
-enum class FrameCookieKind : uint32_t {
+enum class FrameCookieKind : uint8_t {
   Copy,
   XorStackPointer,
   XorFramePointer,
